Render payment buttons from success view
