+++ WooCommerce/Classes/POS/Presentation/CardReaderConnection/UI States/Reader Messages/PointOfSaleCardPresentPaymentSuccessMessageView.swift
@@ -23,6 +23,7 @@ struct PointOfSaleCardPresentPaymentSuccessMessageView: View {
                         .matchedGeometryEffect(id: animation.messageTransitionId, in: animation.namespace, properties: .position)
                 }
             }
+            PaymentsActionButtons()
         }
         .multilineTextAlignment(.center)
     }
