+++ WooCommerce/Classes/POS/Presentation/TotalsView.swift
@@ -16,15 +16,9 @@ struct TotalsView: View {
     private var shouldShowTotalsFields: Bool {
         viewHelper.shouldShowTotalsFields(for: posModel.paymentState)
     }
-    @State private var isShowingPaymentsButtonSpacing: Bool = false
-    @State private var isShowingSendReceiptModal: Bool = false
 
     @Environment(\.dynamicTypeSize) var dynamicTypeSize
     @Environment(\.colorScheme) var colorScheme
-
-    private var shouldShowSendReceiptButton: Bool {
-        ServiceLocator.featureFlagService.isFeatureFlagEnabled(.sendReceiptsForPointOfSale)
-    }
 
     var body: some View {
         HStack {
@@ -48,7 +42,6 @@ struct TotalsView: View {
                                 .transition(.opacity)
                                 .background(cardReaderViewLayout.backgroundColor)
                                 .accessibilityShowsLargeContentViewer()
-                                .minimumScaleFactor(0.1)
                                 .layoutPriority(1)
                         }
 
@@ -61,7 +54,6 @@ struct TotalsView: View {
                         }
                     }
                     .animation(.default, value: posModel.cardPresentPaymentInlineMessage)
-                    paymentsActionButtons
                     Spacer()
                 }
                 .animation(.default, value: isShowingCardReaderStatus)
@@ -78,14 +70,6 @@ struct TotalsView: View {
         }
         .onChange(of: shouldShowTotalsFields, perform: hideTotalsFieldsWithDelay)
         .geometryGroupIfSupported()
-        .posModal(isPresented: $isShowingSendReceiptModal) {
-            POSSendReceiptModalView(sendReceipt: { email in
-                Task { @MainActor in
-                    await posModel.sendReceipt(to: email)
-                }
-            }, isPresented: $isShowingSendReceiptModal)
-            .posModalSizing()
-        }
     }
 
     private var backgroundColor: Color {
@@ -212,62 +196,6 @@ private extension TotalsView {
 }
 
 private extension TotalsView {
-    private var newOrderButton: some View {
-        Button(action: {
-            posModel.startNewCart()
-        }, label: {
-            HStack(spacing: Constants.buttonSpacing) {
-                Text(Localization.newOrder)
-                    .font(Constants.buttonFont)
-            }
-            .frame(minWidth: UIScreen.main.bounds.width / 2)
-        })
-        .padding(Constants.buttonPadding)
-        .foregroundColor(Color.posPrimaryTextInverted)
-        .background(Color.posOverlayFillInverted)
-        .cornerRadius(Constants.buttonCornerRadius)
-    }
-
-    private var sendReceiptButton: some View {
-        Button(action: {
-            isShowingSendReceiptModal = true
-        }, label: {
-            HStack(spacing: Constants.buttonSpacing) {
-                Text(Localization.sendReceipt)
-                    .font(Constants.buttonFont)
-            }
-            .frame(minWidth: UIScreen.main.bounds.width / 2)
-        })
-        .padding(Constants.buttonPadding)
-        .foregroundColor(Color.posPrimaryText)
-        .background(Color.clear)
-        .overlay {
-            RoundedRectangle(cornerRadius: Constants.buttonCornerRadius)
-                        .stroke(Color.posPrimaryText, lineWidth: 1.0)
-        }
-    }
-
-    @ViewBuilder
-    private var paymentsActionButtons: some View {
-        if posModel.paymentState == .cardPaymentSuccessful {
-            if isShowingPaymentsButtonSpacing {
-                Spacer().frame(height: Constants.paymentsButtonSpacing)
-            }
-            sendReceiptButton
-                .renderedIf(shouldShowSendReceiptButton)
-            newOrderButton
-                .onAppear {
-                    isShowingPaymentsButtonSpacing = false
-                    withAnimation(.default.delay(Constants.paymentsButtonButtonSpacingAnimationDelay)) {
-                        isShowingPaymentsButtonSpacing = true
-                    }
-                }
-            Spacer().frame(height: Constants.paymentsButtonSpacing)
-        }
-        else {
-            EmptyView()
-        }
-    }
 
     @ViewBuilder private var cardReaderView: some View {
         switch posModel.cardReaderConnectionStatus {
@@ -362,8 +290,6 @@ private extension TotalsView {
 private extension TotalsView {
     enum Constants {
         static let pricesIdealWidth: CGFloat = 382
-        static let buttonCornerRadius: CGFloat = 8
-
         static let verticalSpacing: CGFloat = 56
 
         static let totalsLineViewPadding: EdgeInsets = .init(top: 20, leading: 24, bottom: 20, trailing: 24)
@@ -380,12 +306,6 @@ private extension TotalsView {
         static let shimmeringWidth: CGFloat = 334
         static let subtotalsShimmeringHeight: CGFloat = 36
         static let totalShimmeringHeight: CGFloat = 40
-
-        static let paymentsButtonSpacing: CGFloat = 80
-        static let paymentsButtonButtonSpacingAnimationDelay: CGFloat = 0.3
-        static let buttonSpacing: CGFloat = 12
-        static let buttonPadding: CGFloat = 32
-        static let buttonFont: POSFontStyle = .posBodyEmphasized
 
         /// Used for synchronizing animations of shimmeringLine and textField
         static let matchedGeometrySubtotalId: String = "pos_totals_view_subtotal_matched_geometry_id"
@@ -408,14 +328,6 @@ private extension TotalsView {
             "pos.totalsView.taxes",
             value: "Taxes",
             comment: "Title for taxes amount field")
-        static let newOrder = NSLocalizedString(
-            "pos.totalsView.newOrder",
-            value: "New order",
-            comment: "Button title for new order button")
-        static let sendReceipt = NSLocalizedString(
-            "pos.totalsView.sendReceipt",
-            value: "Receipt",
-            comment: "Button title for the receipt button")
     }
 }
